Refactor: survey 문항 모듈화, Gender 타입 선언
